Updating ANNDEA Tracker to the optimised version

+++ Code/Utilities/UtilityFunctions.py
@@ -230,7 +230,7 @@ class HitCluster:
            _hit_count=0
            print(TimeStamp(),'Initial number of all possible hit combinations is:',len(_l_Hits)**2)
            print(TimeStamp(),'Number of all possible hit combinations without self-permutations:',(len(_l_Hits)**2)-len(_l_Hits))
-           print(TimeStamp(),'Number of all possible hit  combinations with enforced one-directionality:',((len(_l_Hits)**2)-len(_l_Hits))/2)
+           print(TimeStamp(),'Number of all possible hit  combinations with enforced one-directionality:',int(((len(_l_Hits)**2)-len(_l_Hits))/2))
            for l in _l_Hits:
                _hit_count+=1
                print(TimeStamp(),'Edge generation progress is ',round(100*_hit_count/len(_l_Hits),2), '%',end="\r", flush=True)
